simplify the code
This code has simplified variable names, and the values are directly converted into JSON-formatted string data. It then sends the data to the Antares platform using the send() function.

Note that I also including the projectName and deviceName variables as arguments to the send() function, and we're no longer using the add() function. This is because I already formatting the data as a JSON string, so we don't need to use the add() function to format the data for us. Finally, I also moved the delay statement to the end of the loop() function, so that the code first collects all sensor readings and sends the data, before it waits for the next iteration.